Fixed all three bugs
Sorry this is all in one commit, the changes can be found in the reverse master commit.
Asynced the postMessage function to await the saveMessage call;
used Unshift instead of push after making a copyCopy variable so I don't mutate the original state to force a page rerender and have messages display instantly.
Used a merge sort algorithm to display the messages by time down the page.